Updated bookDetails and associated fragments
ISSUES: The thumbnail is not showing, but the temp is showing on only some of the books. that
makes me believe that the url that im scraping is not being displayed correctly, but is correctly
being scraped. I need to figure out how to show the thumbnail but the rest of the scrape is working